pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frostwatch"
version = "0.1.0"
description = "AI-powered cost and query observability for Snowflake"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "snowflake-connector-python>=3.10",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "anthropic>=0.28",
    "openai>=1.30",
    "google-generativeai>=0.7",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "typer>=0.12",
    "rich>=13",
    "pyyaml>=6",
    "jinja2>=3.1",
    "aiohttp>=3.9",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "types-PyYAML>=6",
]

[project.urls]
Homepage = "https://github.com/arunrajiah/frostwatch"
Repository = "https://github.com/arunrajiah/frostwatch"
Issues = "https://github.com/arunrajiah/frostwatch/issues"

[project.scripts]
frostwatch = "frostwatch.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["frostwatch"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
no_strict_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["frostwatch"]
omit = ["frostwatch/llm/*_provider.py"]
